Moved to remove net46 build target
Also improved the extensions for effects to make it more re-usable

+++ src/OpenRpg.Genres.Fantasy/Defaults/DefaultAttributeStatPopulator.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using OpenRpg.Core.Effects;
 using OpenRpg.Core.Extensions;
 using OpenRpg.Core.Stats;
+using OpenRpg.Genres.Extensions;
 using OpenRpg.Genres.Fantasy.Extensions;
 using OpenRpg.Genres.Fantasy.Stats;
 using OpenRpg.Genres.Fantasy.Types;
@@ -12,13 +13,13 @@ namespace OpenRpg.Genres.Fantasy.Defaults
     {
         public void PopulateStats(IStatsVariables stats, IReadOnlyCollection<Effect> activeEffects)
         {
-            var strengthValue = activeEffects.CalculateBonusFor(EffectTypes.StrengthBonusAmount, EffectTypes.StrengthBonusPercentage);
-            var dexterityValue = activeEffects.CalculateBonusFor(EffectTypes.DexterityBonusAmount, EffectTypes.DexterityBonusPercentage);
-            var constitutionValue = activeEffects.CalculateBonusFor(EffectTypes.ConstitutionBonusAmount, EffectTypes.ConstitutionBonusPercentage);
-            var intelligenceValue = activeEffects.CalculateBonusFor(EffectTypes.IntelligenceBonusAmount, EffectTypes.IntelligenceBonusPercentage);
-            var wisdomValue = activeEffects.CalculateBonusFor(EffectTypes.WisdomBonusAmount, EffectTypes.WisdomBonusPercentage);
-            var charismaValue = activeEffects.CalculateBonusFor(EffectTypes.CharismaBonusAmount, EffectTypes.CharismaBonusPercentage);
-            var allAttributeValue = activeEffects.GetPotencyFor(EffectTypes.AllAttributeBonusAmount);
+            var strengthValue = (int)activeEffects.CalculateBonusFor(EffectTypes.StrengthBonusAmount, EffectTypes.StrengthBonusPercentage);
+            var dexterityValue = (int)activeEffects.CalculateBonusFor(EffectTypes.DexterityBonusAmount, EffectTypes.DexterityBonusPercentage);
+            var constitutionValue = (int)activeEffects.CalculateBonusFor(EffectTypes.ConstitutionBonusAmount, EffectTypes.ConstitutionBonusPercentage);
+            var intelligenceValue = (int)activeEffects.CalculateBonusFor(EffectTypes.IntelligenceBonusAmount, EffectTypes.IntelligenceBonusPercentage);
+            var wisdomValue = (int)activeEffects.CalculateBonusFor(EffectTypes.WisdomBonusAmount, EffectTypes.WisdomBonusPercentage);
+            var charismaValue = (int)activeEffects.CalculateBonusFor(EffectTypes.CharismaBonusAmount, EffectTypes.CharismaBonusPercentage);
+            var allAttributeValue = (int)activeEffects.GetPotencyFor(EffectTypes.AllAttributeBonusAmount);
 
             stats.Strength(strengthValue + allAttributeValue);
             stats.Dexterity(dexterityValue + allAttributeValue);
