+++ src/OpenRpg.Genres.Fantasy/Extensions/EffectExtensions.cs
@@ -10,7 +10,7 @@ namespace OpenRpg.Genres.Fantasy.Extensions
 {
     public static class EffectExtensions
     {
-        public static int CalculateBonusFor(this IReadOnlyCollection<Effect> activeEffects, int amountBonusType, int percentageBonusType, int miscBonus = 0)
+        public static float CalculateBonusFor(this IReadOnlyCollection<Effect> activeEffects, int amountBonusType, int percentageBonusType, int miscBonus = 0)
         {
             var totalAmount = activeEffects.GetPotencyFor(amountBonusType) + miscBonus;
             var totalGenericAmount = activeEffects.GetPotencyFor(EffectTypes.AllAttributeBonusAmount);
